Absorb config crate into types and client
