+++ crates/perfgate-app/Cargo.toml
@@ -17,7 +17,6 @@ categories = ["development-tools"]
 perfgate-types.workspace = true
 perfgate-domain.workspace = true
 perfgate-adapters.workspace = true
-perfgate-config.workspace = true
 perfgate-render.workspace = true
 perfgate-export.workspace = true
 perfgate-sensor.workspace = true
