+++ crates/perfgate-cli/src/main.rs
@@ -27,11 +27,7 @@ use perfgate_app::{
 use perfgate_client::types::{CreateKeyRequest, KeyEntry};
 use perfgate_client::{
     AuthMethod, BaselineClient, ClientConfig, ListBaselinesQuery, ListVerdictsQuery,
-    SubmitVerdictRequest, UploadBaselineRequest,
-};
-use perfgate_config::{
-    ResolvedServerConfig, apply_ratchet_toml_changes, load_config_file,
-    preview_ratchet_toml_changes, resolve_server_config,
+    ResolvedServerConfig, SubmitVerdictRequest, UploadBaselineRequest, resolve_server_config,
 };
 use perfgate_domain::scaling::{
     ScalingReport, SizeMeasurement, classify_complexity, parse_complexity, render_ascii_chart,
@@ -39,6 +35,9 @@ use perfgate_domain::scaling::{
 use perfgate_domain::{DependencyChangeType, SignificancePolicy};
 use perfgate_github::{CommentOptions, GitHubClient};
 use perfgate_render::summary::{SummaryRequest, SummaryUseCase};
+use perfgate_types::config::{
+    apply_ratchet_toml_changes, load_config_file, preview_ratchet_toml_changes,
+};
 use perfgate_types::error::{ConfigValidationError, IoError, PerfgateError};
 use perfgate_types::{
     AggregateWeightMode, AggregationPolicy, BASELINE_REASON_NO_BASELINE, BaselineServerConfig,
