+++ crates/perfgate-client/src/config.rs
@@ -3,8 +3,13 @@
 //! This module defines configuration options for the baseline client,
 //! including authentication, timeouts, and retry behavior.
 
-use std::path::PathBuf;
+use anyhow::Context;
+use perfgate_types::BaselineServerConfig;
+use std::path::{Path, PathBuf};
 use std::time::Duration;
+
+use crate::client::BaselineClient;
+use crate::fallback::FallbackClient;
 
 /// Authentication method for the client.
 #[derive(Debug, Clone, Default)]
@@ -187,6 +192,101 @@ impl ClientConfig {
     }
 }
 
+/// Resolved baseline server configuration with all sources merged.
+#[derive(Debug, Clone, Default)]
+pub struct ResolvedServerConfig {
+    pub url: Option<String>,
+    pub api_key: Option<String>,
+    pub project: Option<String>,
+    pub fallback_to_local: bool,
+}
+
+impl ResolvedServerConfig {
+    /// Returns true if server is configured (has a URL).
+    pub fn is_configured(&self) -> bool {
+        self.url.as_ref().is_some_and(|u| !u.is_empty())
+    }
+
+    /// Creates a [`BaselineClient`] from this configuration.
+    pub fn create_client(&self) -> anyhow::Result<Option<BaselineClient>> {
+        if !self.is_configured() {
+            return Ok(None);
+        }
+
+        let url = self.url.as_ref().expect("checked by is_configured");
+        let mut config = ClientConfig::new(url);
+
+        if let Some(api_key) = &self.api_key {
+            config = config.with_api_key(api_key);
+        }
+
+        let client = BaselineClient::new(config)
+            .with_context(|| format!("Failed to create baseline client for {}", url))?;
+
+        Ok(Some(client))
+    }
+
+    /// Creates a [`FallbackClient`] if fallback is enabled and server is configured.
+    pub fn create_fallback_client(
+        &self,
+        fallback_dir: Option<&Path>,
+    ) -> anyhow::Result<Option<FallbackClient>> {
+        let client = match self.create_client()? {
+            Some(c) => c,
+            None => return Ok(None),
+        };
+
+        let fallback = if self.fallback_to_local {
+            fallback_dir.map(|dir| FallbackStorage::local(dir.to_path_buf()))
+        } else {
+            None
+        };
+
+        Ok(Some(FallbackClient::new(client, fallback)))
+    }
+
+    /// Returns a baseline client for explicit server operations, or an error
+    /// if the server is not configured.
+    pub fn require_client(&self, error_msg: &str) -> anyhow::Result<BaselineClient> {
+        self.create_client()?
+            .ok_or_else(|| anyhow::anyhow!(error_msg.to_string()))
+    }
+
+    /// Returns a baseline client for server operations, or an error if not configured.
+    pub fn require_fallback_client(
+        &self,
+        fallback_dir: Option<&Path>,
+        error_msg: &str,
+    ) -> anyhow::Result<FallbackClient> {
+        self.create_fallback_client(fallback_dir)?
+            .ok_or_else(|| anyhow::anyhow!(error_msg.to_string()))
+    }
+
+    /// Resolve a project for server operations.
+    pub fn resolve_project(&self, project: Option<String>) -> anyhow::Result<String> {
+        project.or_else(|| self.project.clone()).ok_or_else(|| {
+            anyhow::anyhow!(
+                "--project is required (or set --project flag, PERFGATE_PROJECT, or [baseline_server].project in perfgate.toml)"
+            )
+        })
+    }
+}
+
+/// Resolves server configuration from CLI flags, environment variables, and config file values.
+pub fn resolve_server_config(
+    flag_url: Option<String>,
+    flag_key: Option<String>,
+    flag_project: Option<String>,
+    file_config: &BaselineServerConfig,
+) -> ResolvedServerConfig {
+    ResolvedServerConfig {
+        url: flag_url.or_else(|| file_config.resolved_url()),
+        api_key: flag_key.or_else(|| file_config.resolved_api_key()),
+        project: flag_project.or_else(|| file_config.resolved_project()),
+        fallback_to_local: file_config.fallback_to_local,
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -265,5 +365,35 @@ mod tests {
         assert!(matches!(config.auth, AuthMethod::ApiKey(_)));
         assert_eq!(config.timeout, Duration::from_secs(60));
         assert!(config.fallback.is_some());
+    }
+
+    #[test]
+    fn test_resolve_server_config_prefers_flags() {
+        let file_config = BaselineServerConfig {
+            url: Some("https://file.example.com".to_string()),
+            api_key: Some("file-key".to_string()),
+            project: Some("file-project".to_string()),
+            fallback_to_local: false,
+        };
+
+        let resolved = resolve_server_config(
+            Some("https://flag.example.com".to_string()),
+            Some("flag-key".to_string()),
+            Some("flag-project".to_string()),
+            &file_config,
+        );
+
+        assert_eq!(resolved.url.as_deref(), Some("https://flag.example.com"));
+        assert_eq!(resolved.api_key.as_deref(), Some("flag-key"));
+        assert_eq!(resolved.project.as_deref(), Some("flag-project"));
+        assert!(!resolved.fallback_to_local);
+    }
+
+    #[test]
+    fn test_resolved_server_config_reports_unconfigured_without_url() {
+        let resolved = ResolvedServerConfig::default();
+
+        assert!(!resolved.is_configured());
+        assert!(resolved.create_client().unwrap().is_none());
     }
 }
