--- crates/perfgate-config/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "perfgate-config"
-description = "Configuration loading and merging logic for perfgate"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-readme = "README.md"
-documentation = "https://docs.rs/perfgate-config"
-authors.workspace = true
-keywords = ["configuration", "toml", "benchmarking", "perfgate"]
-categories = ["config", "development-tools"]
-
-[dependencies]
-anyhow.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-toml.workspace = true
-toml_edit = "0.23"
-perfgate-types.workspace = true
-perfgate-client.workspace = true
-url.workspace = true
-
-[dev-dependencies]
-tempfile = "3"
-
-[lints]
-workspace = true
