+++ crates/perfgate-types/Cargo.toml
@@ -23,6 +23,7 @@ serde_json.workspace = true
 schemars.workspace = true
 thiserror.workspace = true
 toml.workspace = true
+toml_edit = "0.23"
 arbitrary = { version = "1", features = ["derive"], optional = true }
 
 [dev-dependencies]
