+++ docs/IMPLEMENTATION_PLAN.md
@@ -90,7 +90,6 @@ gate once transition packages stop being publishable.
 |-------|----------------|
 | `perfgate-types` | Core domain types, stable schemas, and fingerprint helpers |
 | `perfgate-api` | API models and authentication types for baseline service |
-| `perfgate-config` | Configuration loading and merging logic |
 | `perfgate-domain` | Core business logic, statistics, budget evaluation, significance, paired analysis, host mismatch logic, and scaling analysis |
 | `perfgate-app` | Orchestration layer for CLI commands |
 | `perfgate-cli` | Command-line interface and argument parsing |
